requirements.txt
aiofiles==0.8.0
backoff==1.11.1
bson==0.5.10
certifi==2022.6.15
charset-normalizer==2.1.0
Deprecated==1.2.13
googleapis-common-protos==1.56.3
grpcio==1.47.0
gunicorn==20.1.0
httptools==0.4.0
idna==3.3
multidict==6.0.2
opentelemetry-api==1.11.1
opentelemetry-exporter-jaeger==1.11.1
opentelemetry-exporter-jaeger-proto-grpc==1.11.1
opentelemetry-exporter-jaeger-thrift==1.11.1
opentelemetry-exporter-otlp==1.11.1
opentelemetry-exporter-otlp-proto-grpc==1.11.1
opentelemetry-exporter-otlp-proto-http==1.11.1
opentelemetry-instrumentation==0.30b1
opentelemetry-instrumentation-aws-lambda==0.30b1
opentelemetry-instrumentation-dbapi==0.30b1
opentelemetry-instrumentation-flask==0.30b1
opentelemetry-instrumentation-grpc==0.30b1
opentelemetry-instrumentation-logging==0.30b1
opentelemetry-instrumentation-pymongo==0.30b1
opentelemetry-instrumentation-requests==0.30b1
opentelemetry-instrumentation-sqlite3==0.30b1
opentelemetry-instrumentation-urllib==0.30b1
opentelemetry-instrumentation-urllib3==0.30b1
opentelemetry-instrumentation-wsgi==0.30b1
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.11.1
opentelemetry-sdk==1.11.1
opentelemetry-semantic-conventions==0.30b1
opentelemetry-util-http==0.30b1
opentracing==2.4.0
protobuf==3.20.1
python-dateutil==2.8.2
requests==2.28.1
sanic==22.6.0
sanic-routing==22.3.0
six==1.16.0
thrift==0.16.0
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.10
uvloop==0.16.0
websockets==10.3
wrapt==1.14.1
